gets a step closer to better project structure - jade plugin still doesn't work :)
