+++ index.js
@@ -1,7 +1,10 @@
-import path from 'path'
+import JadePlugin from './lib/plugins/jade_plugin.js'
+import _ from 'lodash'
 
 module.exports = function (opts) {
-  let spec
-  let outputPath = path.resolve(opts.out)
-  spec.outputPath = outputPath
+  if (!opts) { opts = {} }
+  let spec = { plugins: [new JadePlugin()] }
+  let out = _.defaults(spec, opts)
+
+  return out
 }
