--- lib/app.js
@@ -1,6 +1,0 @@
-var css = require('./assets/css/main.css');
-var index = require('./views/index.jade');
-var about = require('./views/about.jade');
-
-console.log(css);
-console.log(tpl);
