+++ lib/plugins/jade_plugin.js
@@ -1,41 +1,44 @@
 import _ from 'lodash'
-import evaluate from 'eval'
+import cl from 'chalkline'
 
 export default class JadeWebpackPlugin {
 
-  constructor(opts) {
-    if (!opts) { opts = {} };
+  constructor (opts) {
+    if (!opts) { opts = {} }
   }
 
-  apply(compiler) {
-    var self = this;
+  apply (compiler) {
+    // var self = this
 
     compiler.plugin('emit', (compilation, done) => {
       // grab compilation stats
-      let stats = compilation.getStats();
-      let statsJson = stats.toJson();
+      let stats = compilation.getStats()
+      let statsJson = stats.toJson()
 
       // find bundle source
-      let outputFilename = compiler.options.output.filename;
-      let asset = compilation.assets[outputFilename];
+      let outputFilename = compiler.options.output.filename
+      let asset = compilation.assets[outputFilename]
 
-      let source = asset.source();
+      let source = asset.source()
 
       // get jade files
-      let jadeFiles = extractJadeFiles(compiler.records.modules.byIdentifier);
-      console.log(jadeFiles);
+      let jadeFiles = extractJadeFiles(compiler.records.modules.byIdentifier)
 
-      done();
-    });
+      // printe jade files to the console
+      cl.green()
+      console.log(jadeFiles)
+      cl.green()
+      
+      done()
+    })
   }
 
 }
 
 // Utilities
-
-function extractJadeFiles(modules) {
+function extractJadeFiles (modules) {
   return _.compact(_.map(modules, (i, key) => {
-    let filename = key.split('!').pop();
+    let filename = key.split('!').pop()
     if (filename.match(/\.jade$/)) { return filename }
-  }));
+  }))
 }
