+++ package.json
@@ -14,26 +14,26 @@
     "babel-loader": "^6.0.1",
     "babel-preset-es2015": "^6.1.2",
     "babel-runtime": "^6.0.14",
-    "css-loader": "^0.22.0",
     "eval": "^0.1.1",
     "html-webpack-plugin": "^1.6.2",
     "jade": "^1.11.0",
     "jade-html-loader": "0.0.3",
     "jade-loader": "^0.8.0",
-    "lodash": "^3.10.1",
+    "lodash": "^3.10.1"
+  },
+  "devDependencies": {
+    "babel-preset-es2015-node5": "^1.0.0",
+    "chai": "^3.4.1",
+    "chai-fs": "git@github.com:peterkc/chai-fs.git",
+    "chalkline": "0.0.4",
+    "css-loader": "^0.22.0",
     "postcss-cssstats": "^1.0.0",
     "postcss-loader": "^0.7.0",
     "precss": "^1.3.0",
     "rucksack": "^0.1.0",
     "style-loader": "^0.13.0",
-    "webpack": "^1.12.3",
+    "webpack": "^1.12.4",
     "webpack-dev-server": "^1.12.1"
-  },
-  "devDependencies": {
-    "babel-preset-es2015-node5": "^1.0.0",
-    "chai": "^3.4.1",
-    "chalkline": "0.0.4",
-    "webpack": "^1.12.4"
   },
   "license": "MIT",
   "main": "index.js",
@@ -42,8 +42,8 @@
     "url": "git://github.com/carrot/roots-mini.git"
   },
   "scripts": {
+    "start": "./node_modules/.bin/webpack-dev-server",
     "test": "mocha",
-    "watch": "./node_modules/.bin/webpack --watch",
-    "start": "./node_modules/.bin/webpack-dev-server"
+    "watch": "./node_modules/.bin/webpack --watch"
   }
 }
