+++ test/fixtures/jade/views/about.jade
@@ -1,0 +1,4 @@
+extends layout
+
+block content
+  p about page
